// packages/appsync-local-codegen-plugin/src/visitors/appsync-visitor.ts
export interface RawAppSyncLocalConfig extends RawConfig {
  /**
   * @name target
   * @type string
   * @description required, the language target for generated code
   *
   * @example
   * ```yml
   * generates:
   * Models:
   * config:
   *    target: 'swift'
   *  plugins:
   *    - appsync-local-codegen-plugin
   * ```
   * target: 'swift'| 'javascript'| 'typescript' | 'android' | 'metadata'
   */
  target: string;

  /**
   * @name modelName
   * @type string
   * @description optional, name of the model to which the code needs to be generated. Used only
   * when target is set to swift
   * @default undefined, this will generate code for all the models
   *
   * generates:
   * Models:
   * config:
   *    target: 'swift'
   *    model: Todo
   *  plugins:
   *    - appsync-local-codegen-plugin
   * ```
   */
  selectedType: string;

  /**
   * @name metadata
   * @type boolean
   * @description optional, name of the model to which the code needs to be generated
   * @default undefined, this will generate non meta data code
   *
   * generates:
   * Models:
   * config:
   *    target: 'swift'
   *    model: Todo
   *    metadata: true
   *  plugins:
   *    - appsync-local-codegen-plugin
   * ```
   */
  metadata: boolean;
}